src/amisss-core/src/main/java/fr/amisss/core/camunda/CamundaProcessHelper.java
package fr.amisss.core.camunda;

import org.camunda.bpm.engine.RepositoryService;
import org.camunda.bpm.engine.RuntimeService;
import org.camunda.bpm.engine.repository.ProcessDefinition;
import org.camunda.bpm.engine.runtime.ProcessInstance;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

import java.util.List;

@Component
public class CamundaProcessHelper {

    @Autowired
    private RuntimeService runtimeService;

    @Autowired
    private RepositoryService repositoryService;

    public boolean checkProcessActiveByName(final String processDefinitionName) {

        // query for latest process definition with given name
        final ProcessDefinition myProcessDefinition =
                repositoryService.createProcessDefinitionQuery()
                        .processDefinitionName(processDefinitionName)
                        .latestVersion()
                        .singleResult();

        // list all running/unsuspended instances of the process
        final List<ProcessInstance> processInstances =
                runtimeService.createProcessInstanceQuery()
                        .processDefinitionId(myProcessDefinition.getId())
                        .active() // we only want the unsuspended process instances
                        .list();

        return !processInstances.isEmpty();
    }

}
